+++ packages/uploader/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.46 (November 27, 2023)
+
+- fix: add grid background to adjust png 
+
 ## 0.0.45 (十一月 20, 2023)
 
 - chore: set generate image number in dall-e 3